+++ src/main/java/pro/sky/homeworks/homework28/controllers/DepartmentController.java
@@ -7,8 +7,10 @@ import org.springframework.web.bind.annotation.RestController;
 import pro.sky.homeworks.homework28.Employee;
 import pro.sky.homeworks.homework28.services.DepartmentService;
 
+
 import java.util.Collection;
-import java.util.Optional;
+import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/departments")
@@ -21,22 +23,22 @@ public class DepartmentController {
     }
 
     @GetMapping(path = "/max-salary")
-    public Optional<Employee> addEmployee(@RequestParam("departmentId") int departmnetId) {
-        return departmentService.haveMaxSalaryInDept(departmnetId);
+    public Employee addEmployee(@RequestParam("departmentId") int departmentId) {
+        return departmentService.haveMaxSalaryInDept(departmentId);
     }
 
     @GetMapping(path = "/min-salary")
-    public Optional<Employee> removeEmployee(@RequestParam("departmentId") int departmnetId) {
-        return departmentService.haveMinSalaryInDept(departmnetId);
-    }
-
-    @GetMapping(path = "/all")
-    public Collection<Employee> findEmployee(@RequestParam("departmentId") int departmnetId) {
-        return departmentService.getAllByDept(departmnetId);
+    public Employee removeEmployee(@RequestParam("departmentId") int departmentId) {
+        return departmentService.haveMinSalaryInDept(departmentId);
     }
 
     @GetMapping(path = "/al")
-    public Collection<Employee> getEmployees() {
+    public Map<Integer, List<Employee>> getEmployees() {
         return departmentService.getAll();
+    }
+
+    @GetMapping(path = "/all")
+    public Collection<Employee> findEmployee(@RequestParam("departmentId") int departmentId) {
+        return departmentService.getAllByDept(departmentId);
     }
 }
