+++ src/main/java/pro/sky/homeworks/homework28/exceptions/EmployeeNotFoundException.java
@@ -1,10 +1,10 @@
-package pro.sky.homeworks.homework27.exceptions;
+package pro.sky.homeworks.homework28.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
-public class EmployeeNotFoundException extends RuntimeException {
+public class EmployeeNotFoundException extends RuntimeException  {
     public EmployeeNotFoundException(String message) {
         super(message);
     }
