+++ src/main/java/pro/sky/homeworks/homework28/services/DepartmentService.java
@@ -3,29 +3,14 @@ package pro.sky.homeworks.homework28.services;
 import pro.sky.homeworks.homework28.Employee;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.Map;
 
 public interface DepartmentService {
+    Employee haveMaxSalaryInDept(int departmentId);
 
-    Employee addEmployee(Employee employee);
-
-    Employee removeEmployee(Employee employee);
-
-
-    Employee changeSalary(Employee employee, double newSalary);
-
-    Employee changeDepartment(Employee employee, int newDepartment);
-
-    double getSum();
-
-
-    double averageSalary();
-
-    Optional<Employee> haveMaxSalaryInDept(int departmentId);
-
-    Optional<Employee> haveMinSalaryInDept(int departmentId);
+    Employee haveMinSalaryInDept(int departmentId);
 
     List<Employee> getAllByDept(int departmentId);
 
-    List<Employee> getAll();
+    Map<Integer, List<Employee>> getAll();
 }
